Guaruntee Polyselect will not fail
